+++ src/main/java/com/maxmind/geoip2/record/Postal.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This record is returned by all the end points except the Country end point.
  * </p>
  */
-public final class Postal extends AbstractRecord  {
+public final class Postal extends AbstractRecord {
 
     private final String code;
     private final Integer confidence;